Update docs based on feedback @ I/O
